requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.2.0
packaging==24.2
pillow==11.1.0
propcache==0.3.0
PyJWT==2.9.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
twilio==9.5.1
urllib3==2.3.0
yarl==1.18.3
